backend/requirements.txt
# GhostBackup Backend Dependencies

# API server
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Scheduling
APScheduler==3.10.4

# Config / validation
pydantic==2.7.0
PyYAML==6.0.1

# Local SSD sync
xxhash==3.4.1          # ~10 GB/s file hashing for change detection
psutil==5.9.8          # disk usage stats, filesystem type

# Encryption at rest (AES-128-CBC via Fernet — required for accounting data compliance)
cryptography==42.0.5

# File watching (real-time sync)
watchdog==4.0.1

# Utilities
python-multipart==0.0.9

# Testing
pytest>=8.0.0
httpx>=0.27.0          # required by FastAPI TestClient
